Fix incorrect serialization of flag codes
The FlagCode enum was supposed to be serialized as a u16, but
it was being serialized as a u8. This has now been fixed

+++ protocol-v5/src/client.rs
@@ -145,7 +145,7 @@ impl Deserialize for ClientPacket {
 			LOCALPING => match_case!(LocalPing, de),
 			x => {
 				return Err(DeserializeError {
-					ty: DeserializeErrorType::InvalidEnumValue(x),
+					ty: DeserializeErrorType::InvalidEnumValue(x as usize),
 					trace: vec![FieldSpec {
 						field: FieldName::Name("<variant-number>"),
 						ty: "ClientPacket".into(),
